+++ CHANGELOG.md
@@ -5,6 +5,34 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Added
+- Possibility to configure `SchemaDefinitionNamingStrategy` via `configBuilder.forTypesInGeneral().withDefinitionNamingStrategy()`
+- Explicit `DefaultSchemaDefinitionNamingStrategy` that is being applied if no other `SchemaDefinitionNamingStrategy` is being configured
+
+#### Changed
+- Determine names/keys of subschemas in `definitions`/`$defs` through configurable `SchemaDefinitionNamingStrategy`
+
+#### Fixed
+- For multiple definitions for the same type, they might be unnecessarily treated as having conflicting definition names requiring resolution
+
+#### Removed
+- Obsolete `TypeContext.getSchemaDefinitionName()`
+
+## [4.12.2] - 2020-06-10
+### `jsonschema-generator`
+#### Fixed
+- Performance: Cache looked-up getter in `FieldScope.findGetter()`
+- Performance: Cache looked-up field in `MethodScope.findGetterField()`
+
+### `jsonschema-maven-plugin`
+#### Fixed
+- Collecting all classes from a given `<packageName>` even if it contains classes whose super types are not included
+
+## [4.12.1] - 2020-05-28
+### `jsonschema-maven-plugin`
+#### Fixed
+- Setting correct contextual classpath for class/package look-up via reflections
 
 ## [4.12.0] - 2020-05-10
 ### `jsonschema-generator`
@@ -380,7 +408,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
-[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.12.0...HEAD
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.12.2...HEAD
+[4.12.2]: https://github.com/victools/jsonschema-generator/compare/v4.12.1...v4.12.2
+[4.12.1]: https://github.com/victools/jsonschema-generator/compare/v4.12.0...v4.12.1
 [4.12.0]: https://github.com/victools/jsonschema-generator/compare/v4.11.1...v4.12.0
 [4.11.1]: https://github.com/victools/jsonschema-generator/compare/v4.11.0...v4.11.1
 [4.11.0]: https://github.com/victools/jsonschema-generator/compare/v4.10.0...v4.11.0